+++ packages/extender_dart/lib/extender_dart.dart
@@ -26,4 +26,4 @@ export 'src/extensions/universal_extension.dart';
 export 'src/classes/sort_order.dart';
 export 'src/classes/console_text.dart';
 
-export 'src/annotations/class_field_enum.dart';
+export 'src/annotations/field_enum.dart';