+++ packages/extender_dart/lib/src/annotations/field_enum.dart
@@ -1,0 +1,10 @@
+const fieldEnum = FieldEnum._();
+const freezedFieldEnum = FreezedFieldEnum();
+
+class FieldEnum {
+  const FieldEnum._();
+}
+
+class FreezedFieldEnum {
+  const FreezedFieldEnum();
+}