+++ packages/extender_dart_builder/lib/builder.dart
@@ -2,5 +2,9 @@ import 'package:build/build.dart';
 import 'package:extender_dart_builder/class_field_enum/generator.dart';
 import 'package:source_gen/source_gen.dart';
 
-Builder extenderDartBuilder(BuilderOptions options) =>
-    SharedPartBuilder([ClassFieldEnumGenerator()], 'extender_dart');
+import 'freezed_field_enum/generator.dart';
+
+Builder extenderDartBuilder(BuilderOptions options) => SharedPartBuilder(
+      [ClassFieldEnumGenerator(), FreezedFieldEnumGenerator()],
+      'extender_dart',
+    );