+++ packages/extender_dart_builder/lib/extender_dart_builder.dart
@@ -1,3 +1,4 @@
 library extender_dart_builder;
 
 export 'class_field_enum/generator.dart';
+export 'builder.dart';