+++ packages/extender_dart_builder/macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/Users/fukui/fvm/versions/3.10.5"
+export "FLUTTER_ROOT=/Users/fukui/fvm/versions/3.13.9"
 export "FLUTTER_APPLICATION_PATH=/Users/fukui/workspace/project/extender_dart/packages/extender_dart_builder"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_BUILD_DIR=build"