+++ packages/extender_dart/CHANGELOG.md
@@ -9,3 +9,17 @@
 ## 0.0.3
 
 * FEAT: Add uniqueBy method 
+
+## 0.0.4-dev.1
+
+* FEAT: Add annotation
+
+## 0.0.4
+
+* FEAT: Add fieldEnum annotation to generate class field enumerations
+
+## 0.0.5
+* FIX: dependency
+
+## 0.0.6
+* FIX: Add freezedFieldEnum annotation to generate freezed class field