+++ packages/extender_dart_builder/CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* FEAT: class field generator
+* FEAT: class field generator for freezed
+
+## 0.0.2
+
+* FIX: fix warning